edit{api.py, edit_link.py, Gotham_replace,Gotham_link_BDD}: add first version of edit_link and improve some scripts

+++ Orchestrator/Sources/Gotham_link_BDD/remove_in_IDB.py
@@ -38,12 +38,14 @@ def server_in_serv_tag(DB_connection, id="", tag=""):
     cur = DB_connection.cursor()
     # Execute SQL request
     try :
-        if tag=="":
+        if tag=="" and id!="":
             cur.execute("DELETE FROM Serv_Tags WHERE id_serv = ?",(id,))
-        elif id=="":
+        elif id=="" and tag!="":
             cur.execute("DELETE FROM Serv_Tags LEFT JOIN Tags on Serv_Tags.id_tag=Tags.id WHERE tag = ?",(tag,))
-        else :
+        elif id!="" and tag!="":
             cur.execute("DELETE FROM Serv_Tags LEFT JOIN Tags on Serv_Tags.id_tag=Tags.id WHERE id_serv = ? and tag = ?",(id,tag))
+        else:
+            sys.exit(1)
         # Apply the changes
         DB_connection.commit()
         # Logs
@@ -77,13 +79,14 @@ def honeypot_in_hp_tag(DB_connection, id="", tag=""):
     cur = DB_connection.cursor()
     # Execute SQL request
     try :
-        if tag=="":
+        if tag == "" and id != "":
             cur.execute("DELETE FROM Hp_Tags WHERE id_hp = ?",(id,))
-        elif id=="":
+        elif id == "" and tag != "":
             cur.execute("DELETE FROM Hp_Tags LEFT JOIN Tags on Hp_Tags.id_tag=Tags.id WHERE tag = ?",(tag,))
-        else :
+        elif id != "" and tag != "":
             cur.execute("DELETE FROM Hp_Tags LEFT JOIN Tags on Hp_Tags.id_tag=Tags.id WHERE id_hp = ? and tag = ?",(id,tag))
-        
+        else:
+            sys.exit(1)
         # Apply the changes
         DB_connection.commit()
         # Logs
@@ -125,12 +128,20 @@ def link(DB_connection, id):
         logging.error(f"'{id}' removal from the table 'Link' failed : {e}")
         sys.exit(1)
 
-def link_in_link_tags_hp(DB_connection, id):
+def link_in_link_tags_hp(DB_connection, id="", tag_hp=""):
     # Get MariaDB cursor
     cur = DB_connection.cursor()
     # Execute SQL request
     try :
-        cur.execute("DELETE FROM Link_Tags_hp WHERE id_link = ?",(id,))
+        if tag_hp == "" and id != "":
+            cur.execute("DELETE FROM Link_Tags_hp WHERE id_link = ?",(id,))
+        elif id == "" and tag_hp != "":
+            cur.execute("DELETE FROM Link_Tags_hp LEFT JOIN Tags on Link_Tags_hp.id_tag=Tags.id WHERE tag = ?",(tag_hp,))
+        elif id != "" and tag_hp != "":
+            cur.execute("DELETE FROM Link_Tags_hp LEFT JOIN Tags on Link_Tags_hp.id_tag=Tags.id WHERE id_link = ? and tag = ?",(id,tag_hp))
+        else:
+            sys.exit(1)
+        
         # Apply the changes
         DB_connection.commit()
         # Logs
@@ -139,12 +150,20 @@ def link_in_link_tags_hp(DB_connection, id):
         logging.error(f"'{id}' removal from the table 'Link_Tags_hp' failed : {e}")
         sys.exit(1)
 
-def link_in_link_tags_serv(DB_connection, id):
+def link_in_link_tags_serv(DB_connection, id="", tag_serv=""):
     # Get MariaDB cursor
     cur = DB_connection.cursor()
     # Execute SQL request
     try :
-        cur.execute("DELETE FROM Link_Tags_serv WHERE id_link = ?",(id,))
+        if tag_serv == "" and id != "":
+            cur.execute("DELETE FROM Link_Tags_serv WHERE id_link = ?",(id,))
+        elif id == "" and tag_serv != "":
+            cur.execute("DELETE FROM Link_Tags_serv LEFT JOIN Tags on Link_Tags_serv.id_tag=Tags.id WHERE tag = ?",(tag_serv,))
+        elif id != "" and tag_serv != "":
+            cur.execute("DELETE FROM Link_Tags_serv LEFT JOIN Tags on Link_Tags_serv.id_tag=Tags.id WHERE id_link = ? and tag = ?",(id,tag_serv))
+        else:
+            sys.exit(1)
+        
         # Apply the changes
         DB_connection.commit()
         # Logs
